Add tooltips for BlockSettingsView and move <Tooltip /> to shared-hub. Remove some unused translation messages
