+++ _locales/en/messages.json
@@ -1819,7 +1819,7 @@
 		"message": "Ad free"
 	},
 	"ghostery_browser_hub_onboarding_supports_ghosterys_mission": {
-		"message": "Supports Ghostery's Mission"
+		"message": "Supports Ghostery's mission"
 	},
 	"ghostery_browser_hub_onboarding_unlimited_bandwidth": {
 		"message": "Unlimited Bandwidth"
@@ -1861,25 +1861,7 @@
 		"message": "Private by design"
 	},
 	"ghostery_browser_hub_onboarding_private_by_design_description": {
-		"message": "Privacy is incredibly important to us at Ghostery and we've taken extra care to design our products that limits the amount of personal information collected."
-	},
-	"ghostery_browser_hub_onboarding_why_do_you_need_my_email": {
-		"message": "Why do you need my email?"
-	},
-	"ghostery_browser_hub_onboarding_why_do_you_need_my_email_description": {
-		"message": "We've found email is the simplest way to create and secure a unique account without sharing your information. It helps sync your custom settings across your devices."
-	},
-	"ghostery_browser_hub_onboarding_what_do_you_use_my_email_for": {
-		"message": "What do you use my email for?"
-	},
-	"ghostery_browser_hub_onboarding_what_do_you_use_my_email_for_description": {
-		"message": "We never share or sell your personal information. Your email address is used for account management, technical support, product-related communications, and some opt-in features."
-	},
-	"ghostery_browser_hub_onboarding_how_secure_is_": {
-		"message": "How secure is"
-	},
-	"ghostery_browser_hub_onboarding_how_secure_is_ghostery_description": {
-		"message": "We never share or sell your personal information. Your email address is used for account management, technical support, product-related communications, and some opt-in features."
+		"message": "Privacy is incredibly important to us at Ghostery. That’s why we enforce a strict policy when it comes to your data. When you create an account at Ghostery, the only data we collect is your name and email. This information allows us to offer you a user experience wherein you can sync your settings across devices and hear from us about important updates. Under no circumstances do we share or sell this data with any kind of third party."
 	},
 	"ghostery_browser_hub_onboarding_can_i_remove_my_account": {
 		"message": "Can I remove my account?"
@@ -1896,11 +1878,11 @@
 	"ghostery_browser_hub_onboarding_you_are_signed_in_as": {
 		"message": "You are signed in as"
 	},
-	"ghostery_browser_hub_onboarding_yay_youre_all_set": {
-		"message": "Yay! You're all set."
+	"ghostery_browser_hub_onboarding_youve_successfully_set_up_your_browser": {
+		"message": "You've successfully set up your browser!"
 	},
-	"ghostery_browser_hub_onboarding_start_browsing_the_web_with": {
-		"message": "Start browsing the web with"
+	"ghostery_browser_hub_onboarding_surf_with_ease": {
+		"message": "Now you can surf with the greatest of ease."
 	},
 	"ghostery_browser_hub_onboarding_lets_search": {
 		"message": "Let's search"
@@ -1939,7 +1921,7 @@
 		"message": "Blocking \"all trackers\" may cause some websites to break."
 	},
 	"ghostery_browser_hub_info_anti_tracking": {
-		"message": "Anti-tracking anonymizes uniquely identifiable data that trackers try to collect"
+		"message": "Anti-tracking anonymizes uniquely identifiable data that trackers try to collect."
 	},
 	"ghostery_browser_hub_info_smart_browsing": {
 		"message": "Smart-browsing adjusts your blocking settings to decrease page breakage and accelerate page loads."
