+++ app/panel/components/BuildingBlocks/__tests__/CliqzFeature.jsx
@@ -22,7 +22,7 @@ global.t = function(str) {
 };
 
 // Fake the Tooltip implementation
-jest.mock('../../Tooltip');
+jest.mock('../../../../shared-components/Tooltip');
 
 describe('app/panel/components/CliqzFeature.jsx', () => {
 	describe('Snapshot tests with react-test-renderer', () => {
