+++ app/panel/components/BuildingBlocks/__tests__/GhosteryFeature.jsx
@@ -22,7 +22,7 @@ global.t = function(str) {
 };
 
 // Fake the Tooltip implementation
-jest.mock('../../Tooltip');
+jest.mock('../../../../shared-components/Tooltip');
 
 describe('app/panel/components/GhosteryFeature.jsx', () => {
 	describe('Snapshot tests with react-test-renderer', () => {
