+++ app/panel/components/Header.jsx
@@ -15,7 +15,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { ReactSVG } from 'react-svg';
 import ClassNames from 'classnames';
-import Tooltip from './Tooltip';
+import Tooltip from '../../shared-components/Tooltip';
 import HeaderMenu from './HeaderMenu';
 import { sendMessage, sendMessageInPromise } from '../utils/msg';
 import { log } from '../../../src/utils/common';
