+++ app/scss/hub_ghostery_browser.scss
@@ -82,5 +82,4 @@ html, body, #root {
 // Imports from ../shared-components directory
 @import '../shared-components/ToastMessage/ToastMessage.scss';
 @import '../shared-components/ToggleCheckbox/ToggleCheckbox.scss';
-
-
+@import '../shared-components/Tooltip/Tooltip.scss';
