+++ app/scss/panel.scss
@@ -69,7 +69,6 @@ html body {
 @import './partials/_donut_graph';
 @import './partials/_ghostery_feature';
 @import './partials/_cliqz_feature';
-@import './partials/_tooltip';
 @import './partials/_not_scanned';
 @import './partials/_subscribe';
 @import './partials/_stats';
@@ -83,3 +82,4 @@ html body {
 @import '../shared-components/ModalContent/PlusPromoModalContent/PlusPromoModalContent.scss';
 @import '../shared-components/ModalContent/InsightsPromoModalContent/InsightsPromoModalContent.scss';
 @import '../shared-components/ForgotPassword/ForgotPassword.scss';
+@import '../shared-components/Tooltip/Tooltip.scss';
