+++ app/shared-components/Tooltip/Tooltip.scss
@@ -108,6 +108,35 @@
 			line-height: 1.5;
 		}
 	}
+	.onboarding .tooltip-content {
+		color: $white;
+		background-color: $black;
+		font-size: 14px;
+		line-height: 1.5;
+		padding: 11px 10px;
+		width: 150px;
+		text-align: center;
+		position: absolute;
+		z-index: 20;
+		box-shadow: 0px 2px 2px 0 rgba(0, 0, 0, 0.2);
+		box-sizing: border-box;
+		opacity: .96;
+		border-radius: 4px;
+		&:after {
+			box-sizing: border-box;
+			border-style: solid;
+			content: '';
+			position: absolute;
+			border-width: 7px;
+			border-color: transparent transparent $black $black;
+			box-shadow: -2px 2px 2px 0 rgba(0, 0, 0, 0.2);
+			transform-origin: 0 0;
+		}
+		.tooltip-header {
+			font-weight: 500;
+		}
+	}
+
 	.dark-theme .tooltip-content {
 		background-color: #333333;
 		color: #ffffff;
